Add verification user context id with id in command

+++ Modules/VisitSchedule/DentalClinic.VisitSchedule.Application/Command/Handlers/CancellationVisitCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using DentalClinic.VisitSchedule.Core.Exceptions;
+﻿using DentalClinic.Shared.Abstarctions.Services;
+using DentalClinic.VisitSchedule.Core.Exceptions;
 using DentalClinic.VisitSchedule.Core.Repositories;
 using MediatR;
 
@@ -6,10 +7,12 @@ namespace DentalClinic.VisitSchedule.Application.Command.Handlers;
 internal class CancellationVisitCommandHandler : IRequestHandler<CancellationVisitCommand>
 {
     private readonly IVisitRepository _visitRepository;
+    private readonly IUserContextService _userContextService;
 
-    public CancellationVisitCommandHandler(IVisitRepository visitRepository)
+    public CancellationVisitCommandHandler(IVisitRepository visitRepository, IUserContextService userContextService)
     {
         _visitRepository = visitRepository;
+        _userContextService = userContextService;
     }
 
     public async Task<Unit> Handle(CancellationVisitCommand request, CancellationToken cancellationToken)
@@ -18,6 +21,9 @@ internal class CancellationVisitCommandHandler : IRequestHandler<CancellationVis
 
         if (visit == null)
             throw new VisitNotFoundException();
+
+        if (_userContextService.RoleName == "Patient" && (_userContextService.UserId == null || _userContextService.UserId != visit.PatientId))
+            throw new ForbiddenException();
 
         await _visitRepository.DeleteAsync(visit);
